requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
yfinance>=0.2.40
matplotlib>=3.8.0
pyyaml>=6.0.1
joblib>=1.3.2
pytest>=8.0.0